fix: Disable bot on errors instead of throwing an error
Introduced a _config_valid flag to control plugin functionality based on configuration validity. The bot now disables event handling if the configuration is invalid instead of throwing a RunetimeError and revalidates on external config updates. Minor README formatting improvements were also made.
The issue with the error approach was, that it only disabled the bot in the WebUI, but not in reality. When fixing the config or bot, a new bot was started without the old one being ever stopped. These ghost bots still replied to messages and are not stoppable via the WebUI.